requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-classy-tags==0.8.0
django-cms==3.4.5
django-debug-toolbar==1.9.1
django-formtools==2.1
django-google-tag-manager==0.0.3
django-mailer==1.2.2
django-rosetta==0.7.14
django-sekizai==0.10.0
django-treebeard==4.2.0
django-webpack-loader==0.5.0
djangocms-admin-style==1.2.7
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
lockfile==0.12.2
mccabe==0.6.1
microsofttranslator==0.8
polib==1.1.0
psycopg2==2.7.3.2
pylint==1.7.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-memcached==1.58
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
wrapt==1.10.11
Pillow
easy-thumbnails
django-filer
django-modeltranslation
djangocms-text-ckeditor
django-el-pagination
